Refactoring

+++ api/Configuration.go
@@ -9,19 +9,20 @@ const (
 	CONFIG_FILE = "hub-config.json"
 )
 
+// Configuration for Mail when sending out email (e.g. rule triggers)
 type MailConfig struct {
 	User string
 	Pass string
 	Host string
-	Port int
+	Port uint16
 }
 
 type Configuration struct {
 	Name       string
 	NodeID     string
 	Db         string
-	MaxProcs   int
-	ServerPort int
+	MaxProcs   uint8
+	ServerPort uint16
 	Protocols  map[string]ProtocolConfiguration
 	Mail       MailConfig
 }
